+++ src/components/screen/screen-list.jsx
@@ -1,6 +1,5 @@
 import { React, useEffect, useState, useContext } from "react";
 import { useTranslation } from "react-i18next";
-import { useDispatch } from "react-redux";
 import UserContext from "../../context/user-context";
 import List from "../util/list/list";
 import { ScreenColumns } from "./util/screen-columns";
@@ -13,7 +12,6 @@ import {
   useGetV2ScreensQuery,
   useDeleteV2ScreensByIdMutation,
   useGetV2ScreensByIdScreenGroupsQuery,
-  api,
 } from "../../redux/api/api.generated.ts";
 import {
   displaySuccess,